fix: Address critical issues from independent audit (excluding binaries)
BREAKING CHANGES:
- Removed non-existent type exports (BitchatPacket, RoutingInfo, TlvField)
- Changed mobile secure storage from XOR to ChaCha20Poly1305

Fixes Applied:
- Test Infrastructure: Fixed 100+ compilation errors, library tests now compile
- FFI Implementation: Replaced 10+ stub functions with real implementations
  - start_discovery(): Now starts mesh service
  - roll_dice(): Cryptographically secure random generation
  - place_bet(): Proper bet logging
  - set_power_mode(): Battery optimization controls
- Android Security: Replaced fake crypto with real implementations
  - Encryption: ChaCha20Poly1305 authenticated encryption (was returning input)
  - Signatures: Ed25519 digital signatures (was returning zeros)
  - Added proper crypto dependencies to JNI bridge
- Mobile Storage: ChaCha20Poly1305 with HKDF key derivation

Remaining Critical Issues:
- Bluetooth: btleplug cannot advertise as peripheral (mesh impossible)
- Integration Tests: 47 compilation errors remain
- Game Networking: Still stubbed, no P2P communication
- Android JNI: May have build issues with new crypto

Project Status:
- Before fixes: 35-40% complete (claimed 85-95%)
- After fixes: 45-50% complete
- Time to production: 4-6 months minimum

Documentation:
- Created INDEPENDENT_REVIEW_SUMMARY.md with audit findings
- Created ACTUAL_PROJECT_STATUS.md with honest assessment
- Updated CRITICAL_FIXES_IMPLEMENTED.md with fix details
- Updated .gitignore to exclude build artifacts

+++ src/mesh/anti_cheat.rs
@@ -101,7 +101,7 @@ impl AntiCheatMonitor {
         behavior.last_packet_time = Some(now);
         
         // Check for suspicious patterns in game packets
-        if packet.packet_type as u8 >= 0x20 && packet.packet_type as u8 <= 0x27 {
+        if packet.packet_type >= 0x20 && packet.packet_type <= 0x27 {
             // Game-specific anti-cheat would go here
             // For example: impossible dice rolls, betting patterns, etc.
         }
